DB migration to drop table ItemCategory and add table Tag
We don't need to create table ItemCategory in db version 1 for new
users, as it will be created as Tag table in migration 4. For users who
already have the table ItemCategory, the table is empty so it is safe to
drop it. From migration 4, we will start using table Tag